requirements.txt
matplotlib==3.2
numpy==1.22.4
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
Pillow>=9.5.0
PyAutoGUI==0.9.54
